src/main/java/br/com/iteris/universidade/testes/intro/controller/PersonController.java
package br.com.iteris.universidade.testes.intro.controller;

import br.com.iteris.universidade.testes.intro.model.Person;
import br.com.iteris.universidade.testes.intro.service.PersonService;
import lombok.RequiredArgsConstructor;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping(value = "/person", produces = MediaType.APPLICATION_JSON_VALUE, consumes = MediaType.APPLICATION_JSON_VALUE)
@RequiredArgsConstructor
public class PersonController {
    private final PersonService personService;

    @PostMapping
    public ResponseEntity<Person> insert(@RequestBody Person person){
        return ResponseEntity.ok(personService.insert(person));
    }
}
